Refactor if domain objects into their own class files
