+++ test/unit/uk/gov/hmrc/individualsincomeapi/domain/integrationframework/paye/GrossEarningsForNicsSpec.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package unit.uk.gov.hmrc.individualsincomeapi.domain.integrationframework
+package unit.uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye
 
 import play.api.libs.json.Json
-import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.GrossEarningsForNics
+import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye.GrossEarningsForNics
+import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye.IncomePaye._
 import utils.UnitSpec
-import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.IncomePaye._
 
 class GrossEarningsForNicsSpec extends UnitSpec {
